Add errors blueprint
