+++ src/databass/errors/templates/error.html
@@ -1,0 +1,7 @@
+{% extends '_base.html' %}
+{% block content %}
+    <h1>Application Error</h1>
+    <p>{{ error }}</p>
+    <a href="{{ back }}">back</a>
+    <a href="/src/databass/static">home</a>
+{% endblock %}
