+++ src/databass/templates/error.html
@@ -3,5 +3,5 @@
     <h1>Application Error</h1>
     <p>{{ error }}</p>
     <a href="{{ back }}">back</a>
-    <a href="/">home</a>
+    <a href="/src/databass/static">home</a>
 {% endblock %}
